+++ crates/re_sdk_python/src/python_bridge.rs
@@ -94,6 +94,7 @@ fn rerun_sdk(py: Python<'_>, m: &PyModule) -> PyResult<()> {
     m.add_function(wrap_pyfunction!(log_rects, m)?)?;
 
     m.add_function(wrap_pyfunction!(log_camera, m)?)?;
+    m.add_function(wrap_pyfunction!(log_point, m)?)?;
     m.add_function(wrap_pyfunction!(log_points, m)?)?;
     m.add_function(wrap_pyfunction!(log_path, m)?)?;
     m.add_function(wrap_pyfunction!(log_line_segments, m)?)?;
@@ -612,6 +613,77 @@ fn log_rects(
 }
 
 // ----------------------------------------------------------------------------
+
+/// Log a 2D or 3D point.
+///
+/// `position` is either 2x1 or 3x1.
+///
+/// If no `space` is given, the space name "2D" or "3D" will be used,
+/// depending on the dimensionality of the data.
+#[pyfunction]
+fn log_point(
+    obj_path: &str,
+    position: numpy::PyReadonlyArray1<'_, f32>,
+    color: Option<Vec<u8>>,
+    timeless: bool,
+    space: Option<String>,
+) -> PyResult<()> {
+    let dim = match position.shape() {
+        [2] => 2,
+        [3] => 3,
+        shape => {
+            return Err(PyTypeError::new_err(format!(
+                "Expected either a 2D or 3D position; got {shape:?}"
+            )));
+        }
+    };
+
+    let mut sdk = Sdk::global();
+
+    let obj_path = parse_obj_path(obj_path)?;
+
+    sdk.register_type(
+        obj_path.obj_type_path(),
+        if dim == 2 {
+            ObjectType::Point2D
+        } else {
+            ObjectType::Point3D
+        },
+    );
+
+    let time_point = time(timeless);
+
+    if let Some(color) = color {
+        let color = convert_color(color)?;
+        sdk.send_data(
+            &time_point,
+            (&obj_path, "color"),
+            LoggedData::Single(Data::Color(color)),
+        );
+    }
+
+    let position = position.as_slice()?;
+    let pos_data = match dim {
+        2 => Data::Vec2([position[0], position[1]]),
+        3 => Data::Vec3([position[0], position[1], position[2]]),
+        _ => unreachable!(),
+    };
+
+    sdk.send_data(
+        &time_point,
+        (&obj_path, "pos"),
+        LoggedData::Single(pos_data),
+    );
+
+    let space = space.unwrap_or_else(|| if dim == 2 { "2D" } else { "3D" }.to_owned());
+    sdk.send_data(
+        &time_point,
+        (&obj_path, "space"),
+        LoggedData::Single(Data::Space(parse_obj_path(&space)?)),
+    );
+
+    Ok(())
+}
 
 /// positions: Nx2 or Nx3 array
 /// * `colors.len() == 0`: no colors